pyproject.toml
[project]
name = "dog-trial-form-evaluation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.38.3",
    "coverage>=7.8.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "starlette>=0.46.2",
    "streamlit>=1.44.1",
    "uvicorn>=0.34.2",
]
